narrow qmux public API
